+++ src/main/java/com/kitchenhack/apikitchen/dtos/SuscripcionPlanDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 public class SuscripcionPlanDTO {
 
     private Integer id;
-    private Long idUsuario;       // FK a usuario (Long)
+    private Integer idUsuario;       // FK a usuario (Integer)
     private Integer idPlan;       // FK a plan_maestro (Integer)
     private LocalDate fechaInicio; // Se asigna automáticamente en el servidor
     private Boolean activo;        // Siempre TRUE al crear
@@ -14,8 +14,8 @@ public class SuscripcionPlanDTO {
     public Integer getId() { return id; }
     public void setId(Integer id) { this.id = id; }
 
-    public Long getIdUsuario() { return idUsuario; }
-    public void setIdUsuario(Long idUsuario) { this.idUsuario = idUsuario; }
+    public Integer getIdUsuario() { return idUsuario; }
+    public void setIdUsuario(Integer idUsuario) { this.idUsuario = idUsuario; }
 
     public Integer getIdPlan() { return idPlan; }
     public void setIdPlan(Integer idPlan) { this.idPlan = idPlan; }
