+++ src/main/java/com/kitchenhack/apikitchen/entities/Recipe.java
@@ -10,7 +10,7 @@ public class Recipe {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "titulo", nullable = false)
     private String title;
@@ -45,7 +45,7 @@ public class Recipe {
     public Recipe() {
     }
 
-    public Recipe(Long id, String title, String description, Usuario idAutor, Integer prepTimeMinutes, String difficulty, Long categoryId, Double totalCalories, Double averageRating, Boolean published, LocalDateTime ultimaActualizacion) {
+    public Recipe(Integer id, String title, String description, Usuario idAutor, Integer prepTimeMinutes, String difficulty, Long categoryId, Double totalCalories, Double averageRating, Boolean published, LocalDateTime ultimaActualizacion) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -56,11 +56,11 @@ public class Recipe {
         this.ultimaActualizacion = ultimaActualizacion;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
