+++ src/main/java/com/kitchenhack/apikitchen/repositories/ProgresoSaludRepository.java
@@ -10,9 +10,8 @@ import java.util.Optional;
 
 @Repository
 public interface ProgresoSaludRepository extends JpaRepository<ProgresoSalud, Integer> {
-    Optional<ProgresoSalud> findByIdUsuario_Id(Long usuarioId);
+    Optional<ProgresoSalud> findByIdUsuario_Id(Integer usuarioId);
 
     // Filtra mediciones de un usuario en un rango de fechas (US-P4-S2-03)
-    List<ProgresoSalud> findByIdUsuario_IdAndFechaBetween(Long usuarioId, LocalDate inicio, LocalDate fin);
+    List<ProgresoSalud> findByIdUsuario_IdAndFechaBetween(Integer usuarioId, LocalDate inicio, LocalDate fin);
 }
-
